+++ AppVersion.cs
@@ -1,13 +1,13 @@
-namespace MissileCamera
+﻿namespace MissileCamera
 {
     internal static class AppVersion
     {
-        public const string ReleaseBase = "0.27.3";
+        public const string ReleaseBase = "0.27.4";
         public const string BepInSemVer = ReleaseBase;
-        public const string VersionChannel = "DEV";
+        public const string VersionChannel = "REL";
         public const int CycleBuildNumber = 1;
-        public const string ChangeLetters = "P";
-        public const int SubNumber = 5;
-        public const string DisplayVersion = "0.27.3 Build DEV1P5";
+        public const string ChangeLetters = "";
+        public const int SubNumber = 0;
+        public const string DisplayVersion = ReleaseBase;
     }
 }